+++ libs/small-steps/src/Control/State/Transition/Extended.hs
@@ -56,9 +56,11 @@ module Control.State.Transition.Extended (
   failOnJust,
   failOnNonEmpty,
   failOnNonEmptySet,
+  failOnNonEmptyMap,
   failureOnJust,
   failureOnNonEmpty,
   failureOnNonEmptySet,
+  failureOnNonEmptyMap,
   judgmentContext,
   trans,
   liftSTS,
@@ -109,6 +111,9 @@ import Data.Functor (($>), (<&>))
 import Data.Kind (Type)
 import Data.List.NonEmpty (NonEmpty (..))
 import qualified Data.List.NonEmpty as NE
+import Data.Map.NonEmpty (NonEmptyMap)
+import qualified Data.Map.NonEmpty as NEM
+import Data.Map.Strict (Map)
 import Data.Set.NonEmpty (NonEmptySet)
 import qualified Data.Set.NonEmpty as NES
 import Data.Typeable (typeRep)
@@ -430,6 +435,16 @@ failOnNonEmptySet cond onNonEmpty = liftF $ Predicate (failureOnNonEmptySet cond
 failureOnNonEmptySet ::
   (Foldable f, Ord a) => f a -> (NonEmptySet a -> e) -> Validation (NonEmpty e) ()
 failureOnNonEmptySet cond = failureOnJust (NES.fromFoldable cond)
+
+-- | Produce a predicate failure when supplied foldable is not an empty Map, contents of which
+-- will be converted to a NonEmptyMap and can be used inside the predicate failure.
+failOnNonEmptyMap :: Map k v -> (NonEmptyMap k v -> PredicateFailure sts) -> Rule sts ctx ()
+failOnNonEmptyMap cond onNonEmpty = liftF $ Predicate (failureOnNonEmptyMap cond onNonEmpty) id ()
+
+-- | Produce a failure when supplied foldable is not an empty Map, contents of which will be
+-- converted to a NonEmptyMap and can then be used inside the failure constructor.
+failureOnNonEmptyMap :: Map k v -> (NonEmptyMap k v -> e) -> Validation (NonEmpty e) ()
+failureOnNonEmptyMap cond = failureOnJust (NEM.fromMap cond)
 
 -- | Oh noes with an explanation
 --